build: Separate lib from bin
